lots of style changes and unused variable commenting, starting to sink into some meatier content changes but want to commit first.

+++ library/src/G4CMPSCUtils.cc
@@ -340,7 +340,7 @@ GetTauAsAFunctionOfEnergy( const std::vector<std::vector<G4double> > & tauVsPhon
 
     //If our bin guess is the final bin, then just use that bin value. We'll
     //accept a single edge case here.
-    if (binGuess == tauVsPhononEnergy.size()-1) {
+    if (binGuess == ((int)tauVsPhononEnergy.size())-1) {
       return tauVsPhononEnergy[binGuess][1];
     } else {
       //^Otherwise, try to interpolate.
